refactor: Extract JSL step conversion to dedicated adapter
